+++ Changelog.md
@@ -39,6 +39,7 @@
 - Combine 'Split Bodies', 'Flip Winding Order', and 'Export' into a 'Mesh' submenu
 - Correctly calculate volume with non-uniform layer height
 - Fix the way paths are inflated in the PCB tool
+- Allow configuring max allowed circle error in PCB tool
 
 ## v0.9.2 &mdash; August 30th, 2026